Labs/RAG_QandA/app.py
import gradio as gr
import pandas as pd
from langchain_community.document_loaders.csv_loader import CSVLoader
from langchain_openai import ChatOpenAI, OpenAIEmbeddings
from langchain_community.docstore.in_memory import InMemoryDocstore
from langchain_community.vectorstores import FAISS
from langchain_core.prompts import ChatPromptTemplate
from langchain.chains import create_retrieval_chain
from langchain.chains.combine_documents import create_stuff_documents_chain
import faiss
import os

# Set your OpenAI API key here or use a secret in HF Spaces
os.environ["OPENAI_API_KEY"] = os.getenv("OPENAI_API_KEY", "")

# Build the RAG pipeline from a CSV file
def build_rag(csv_path):
    # ##################################
    # # INSERT YOUR CODE HERE
    # loader = 
    # docs = 
    # embeddings = 
    # index = 
    # vector_store = 
    # retriever = 
    # llm = 
    # system_prompt = 
    # prompt = 
    # question_answer_chain = 
    # rag_chain = 
    # ##################################
    return rag_chain

rag_chain = None

def upload_csv(file):
    global rag_chain
    rag_chain = build_rag(file.name)
    return "CSV loaded and indexed! You can now ask questions."

def ask_question(question):
    if rag_chain is None:
        return "Please upload a CSV file first."
    try:
        answer = rag_chain.invoke({"input": question})
        return answer['answer']
    except Exception as e:
        return f"Error: {e}"

with gr.Blocks() as demo:
    gr.Markdown("# Simple CSV RAG Q&A Bot\nUpload a CSV and ask questions about its content!")
    with gr.Row():
        csv_input = gr.File(label="Upload CSV", file_types=[".csv"])
        upload_btn = gr.Button("Load CSV")
    status = gr.Textbox(label="Status", interactive=False)
    question = gr.Textbox(label="Ask a question about your CSV")
    answer = gr.Textbox(label="Answer", interactive=False)
    ask_btn = gr.Button("Ask")

    upload_btn.click(upload_csv, inputs=csv_input, outputs=status)
    ask_btn.click(ask_question, inputs=question, outputs=answer)

demo.launch()
