Labs/RAG_QandA/requirements.txt
gradio
pandas
langchain==0.1.14
langchain-community
langchain-openai
faiss-cpu
python-dotenv
openai
